Moved private functions to submodule and improved docstring
