+++ src/worldcereal/openeo/mapping.py
@@ -1,0 +1,129 @@
+"""Private methods for cropland/croptype mapping. The public functions that
+are interracting with the methods here are defined in the `worldcereal.job`
+sub-module.
+"""
+from typing import TYPE_CHECKING
+
+from openeo import DataCube
+from openeo_gfmap.features.feature_extractor import apply_feature_extractor
+from openeo_gfmap.inference.model_inference import apply_model_inference
+from openeo_gfmap.preprocessing.scaling import compress_uint8, compress_uint16
+
+if TYPE_CHECKING:  # To avoid cyclic imports during runtime
+    from worldcereal.job import CropLandParameters, CropTypeParameters
+
+
+def _cropland_map(
+    inputs: DataCube, cropland_parameters: "CropLandParameters"
+) -> DataCube:
+    """Method to produce cropland map from preprocessed inputs, using
+    a Presto feature extractor and a CatBoost classifier.
+
+    Parameters
+    ----------
+    inputs : DataCube
+        preprocessed input cube
+
+    Returns
+    -------
+    DataCube
+        binary labels and probability
+    """
+
+    # Run feature computer
+    features = apply_feature_extractor(
+        feature_extractor_class=cropland_parameters.feature_extractor,
+        cube=inputs,
+        parameters=cropland_parameters.features_parameters.dict(),
+        size=[
+            {"dimension": "x", "unit": "px", "value": 100},
+            {"dimension": "y", "unit": "px", "value": 100},
+        ],
+        overlap=[
+            {"dimension": "x", "unit": "px", "value": 0},
+            {"dimension": "y", "unit": "px", "value": 0},
+        ],
+    )
+
+    # Run model inference on features
+    classes = apply_model_inference(
+        model_inference_class=cropland_parameters.classifier,
+        cube=features,
+        parameters=cropland_parameters.classifier_parameters.dict(),
+        size=[
+            {"dimension": "x", "unit": "px", "value": 100},
+            {"dimension": "y", "unit": "px", "value": 100},
+            {"dimension": "t", "value": "P1D"},
+        ],
+        overlap=[
+            {"dimension": "x", "unit": "px", "value": 0},
+            {"dimension": "y", "unit": "px", "value": 0},
+        ],
+    )
+
+    # Cast to uint8
+    classes = compress_uint8(classes)
+
+    return classes
+
+
+def _croptype_map(
+    inputs: DataCube,
+    croptype_parameters: "CropTypeParameters",
+    cropland_mask: DataCube = None,
+) -> DataCube:
+    """Method to produce croptype map from preprocessed inputs, using
+    a Presto feature extractor and a CatBoost classifier.
+
+    Parameters
+    ----------
+    inputs : DataCube
+        preprocessed input cube
+    cropland_mask : DataCube, optional
+        optional cropland mask, by default None
+
+    Returns
+    -------
+    DataCube
+        croptype labels and probability
+    """
+
+    # Run feature computer
+    features = apply_feature_extractor(
+        feature_extractor_class=croptype_parameters.feature_extractor,
+        cube=inputs,
+        parameters=croptype_parameters.feature_parameters.dict(),
+        size=[
+            {"dimension": "x", "unit": "px", "value": 100},
+            {"dimension": "y", "unit": "px", "value": 100},
+        ],
+        overlap=[
+            {"dimension": "x", "unit": "px", "value": 0},
+            {"dimension": "y", "unit": "px", "value": 0},
+        ],
+    )
+
+    # Run model inference on features
+    classes = apply_model_inference(
+        model_inference_class=croptype_parameters.classifier,
+        cube=features,
+        parameters=croptype_parameters.classifier_parameters.dict(),
+        size=[
+            {"dimension": "x", "unit": "px", "value": 100},
+            {"dimension": "y", "unit": "px", "value": 100},
+            {"dimension": "t", "value": "P1D"},
+        ],
+        overlap=[
+            {"dimension": "x", "unit": "px", "value": 0},
+            {"dimension": "y", "unit": "px", "value": 0},
+        ],
+    )
+
+    # Mask cropland
+    if cropland_mask is not None:
+        classes = classes.mask(cropland_mask == 0, replacement=0)
+
+    # Cast to uint16
+    classes = compress_uint16(classes)
+
+    return classes
